fix: rename shadowing `io` to `io_`
Closes #901.
