+++ docs/quick_start.md
@@ -101,7 +101,7 @@ In this case, we consider the [Hello World agent service](./hello_world_agent_se
 
 2. Use the CLI to download and build the images to deploy the [Hello World agent service](./hello_world_agent_service.md):
     ```bash
-    autonomy deploy build deployment valory/hello_world:0.1.0:bafybeibjvvft6ojpoiue3mj5t3f7cm42kez6q4yy4rsrbhulzzruziy5na keys.json --remote
+    autonomy deploy build deployment valory/hello_world:0.1.0:bafybeia7lxamdmgdc2otag2g7oloifx2y75nlhsltjmfwm2riqzva56bjm keys.json --remote
     ```
     The command above generates the required images to run the agent service using the keys provided in the `keys.json` file. In this case, we are accessing the service definition located in the Service Registry.
