+++ packages/hashes.csv
@@ -4,16 +4,16 @@ open_aea/protocols/scaffold,bafybeihhyjuqvp5zqtikkjrwtikg4fr4tsub43n777huxtpeu77
 open_aea/protocols/signing,bafybeid6f5ceool4evgaxdegs5pkjoivqjk3mo3ehctalswngb5m5c32wm
 open_aea/skills/scaffold,bafybeih3h7yia76rrbittnoeotpraiuho5rdqsnksuvncinaa3akgmnbaa
 valory/agents/abstract_abci,bafybeibw3zlvxwj5y3tgd7zreet4tpoyqv2kgefbezqtq5j7r6765yrsz4
-valory/agents/apy_estimation,bafybeiep4mcz5n4r2uu4rco2gf4ouxptnmd6wt2ais6gbjmpteweziq4wy
-valory/agents/apy_estimation_chained,bafybeifzm5xat4uakcax643hf5g5cma2scyeubzrdrtbepmgi4ujvulyxe
+valory/agents/apy_estimation,bafybeihfufuc7gmkl3hvncgfaknx4xgz6klzigxrmm5jzbtincdfvw3bju
+valory/agents/apy_estimation_chained,bafybeibomuupinknuohcenw3s2n7pnhkwrgdopaz3g3tnbuqawbbdsx2vy
 valory/agents/counter,bafybeigtvkbkpdrt2hra47zdt72mbbmp653plwf7te43lovij4yp4vqp6q
 valory/agents/counter_client,bafybeicseyonjog4h53kobzehbm7zf53qgww3kdhcuo57bp42kjienwpeq
-valory/agents/hello_world,bafybeibotci3wyi6nyduztw6nhcosgci7brs2z3uzvs3fph7xxmcjm2xje
-valory/agents/oracle,bafybeiexc53bm7af643wcbdwa47dipfim3lapymru2emw76b3ayv4baxma
-valory/agents/register_reset,bafybeiguas2i2xjsh5ngtkgfofzng4tyjonxhvxplzqdozgm6votmk4ape
-valory/agents/registration_start_up,bafybeihrf4qsbn63xk4deeouxk7ttar2s4epdlszl6nh3yhoqx2mrotnua
-valory/agents/simple_abci,bafybeiei4nte44ryfyjbpahimzir4xtepqss7x42kcq5iljzg5vsesnzi4
-valory/agents/test_abci,bafybeiaj2w7zn4kw4ka6hoq4gmw3zggvb2fvm2wg7znhmnz32cjfk67e5y
+valory/agents/hello_world,bafybeigjyz4pd7ylyxxpryadz7y5q7idxxamm7xh6rh2gtqz2tt276ii4e
+valory/agents/oracle,bafybeih2qdphkai3flbuywwmwmghkcbsfqephlszfa7vdbifnwtzm3a7ne
+valory/agents/register_reset,bafybeicdbjj7gknew3efdvus64w4rq33mns4qzipejwjvqxpd4jculxkpm
+valory/agents/registration_start_up,bafybeia7bsali3v55shtq3burzknhqjcjxiyopvrbshcaqesvcxj57bd2q
+valory/agents/simple_abci,bafybeieuog75diggn7kxa757nyz3jyhjs7qy75pmzvfxc2vi35b6qa3hcq
+valory/agents/test_abci,bafybeicrcjw7tjwtltqvtctfgdyodm2zi4n3tvxeususbqzlpcvzy5e2gq
 valory/connections/abci,bafybeiaau6gvge6gh5xp7brrpljqe3mhosy3p67e5cgivtypj7nplmii24
 valory/connections/http_client,bafybeiajlo43jrfdhtkpagjtj4izm2rc24ysbnzbtxnu5qqc2pxoov66z4
 valory/connections/ledger,bafybeiemp6u7xsyvwcq25i3t62oiykwsg5wmp7xs2e6eidll5vf3htdpwy
@@ -31,33 +31,33 @@ valory/protocols/contract_api,bafybeigkedpmqhm7lrqusnkyfcapyx3dlkz2djfq74hou5pwt
 valory/protocols/http,bafybeienmids6f5pcey2bnywmiup2faawqgmdhhz65qfciuhkntugztvpi
 valory/protocols/ledger_api,bafybeif2c3jt4tim3zbfq5c2lds7bs6qj5gvj2obk3pzdk3uy6zk2ivo3e
 valory/protocols/tendermint,bafybeihcnjhovvyyfbkuw5sjyfx2lfd4soeocfqzxz54g67333m6nk5gxq
-valory/services/apy_estimation,bafybeie74yw6dfddn7yfsfvsktr7ea5vt7u26un6iwam6oxd65norcdg4m
-valory/services/apy_estimation_hardhat,bafybeibcvg26tapbeyrmmvj4sfdgkhefjxeaqiemv7vjnkwzwbyfwroptq
+valory/services/apy_estimation,bafybeifvmiaymtyi2ecrxbijul24vsftynckwect4f55kgtreuga4nekbi
+valory/services/apy_estimation_hardhat,bafybeihvhr76k4a2clipajs5to54cntzybedlxdxxt7nawd3woc25mdtfy
 valory/services/counter,bafybeicgcqytl2hoy67hmmi5vbhq2sjzkj4d5eajudjav3d6aqskl2bsge
-valory/services/hello_world,bafybeibjvvft6ojpoiue3mj5t3f7cm42kez6q4yy4rsrbhulzzruziy5na
-valory/services/oracle_goerli,bafybeiby4qy7aebbqxou6n5k5hnstyfxehzbtd66fp3tunuxw4q5l5beli
-valory/services/oracle_hardhat,bafybeigkz3nrzvj2mouyrtv2dhemdv44tahuuvbuhlng2us5t64omlbusq
-valory/services/oracle_polygon,bafybeie27m3rh4tbbueolzyeq3frvtcjdmoqme4jdxfcmfxutla6zg4ijm
-valory/services/oracle_ropsten,bafybeiafaeocntdta375gz56wq2u6c5q2wd3u2vrdizgf6pearmtgizev4
-valory/services/oracle_ropsten_cluster,bafybeia56rmgt6wwg33aqtwp5frt24fl6vqq25npf2ui5msxtqbkfyxs34
-valory/services/register_reset,bafybeig32fkllhkblkexhgfttshnlpzdn5vgukbccr2thn7mq7d5fgkppu
-valory/services/simple_abci,bafybeicyg4d4f35dn3uinr46qm3josvksqp6onpf7kyfq36xlpprmun57a
+valory/services/hello_world,bafybeia7lxamdmgdc2otag2g7oloifx2y75nlhsltjmfwm2riqzva56bjm
+valory/services/oracle_goerli,bafybeiaer5lepyb6ryarcwhozfqg63ctzgzmge4ghrjwya53lxzadswzbu
+valory/services/oracle_hardhat,bafybeihq2rfeoz6yfw6skvblmk4jwpv7xvga6cu5uou5z7h3b5objaxzlq
+valory/services/oracle_polygon,bafybeicng2tuffqgxd7zqjd2onudo5cpaunjddx5i75tv47aia3g245omq
+valory/services/oracle_ropsten,bafybeigtlbyv6zjnbh4apdmjq6iwf3n4lycmyzi6nimnky55s5sl44ww5e
+valory/services/oracle_ropsten_cluster,bafybeidpjae2ljw6hua2ho3xewgsdvlppkeb4jjjvfffj53lywmydpe2a4
+valory/services/register_reset,bafybeiefpdsggmrn2yx2ntgyd7kpvbubspfl4dwyaetj4wxbzh77wwbo5i
+valory/services/simple_abci,bafybeia3kbuxiukvjejyug2uxnzskamv4wge736yu5xqtnrfqimztoymde
 valory/skills/abstract_abci,bafybeiam52ztcss46i3tgywuv4qmvddqgn7exmbt5kcicqpyip4kfeyap4
-valory/skills/abstract_round_abci,bafybeie4ux473o45jjo5cb2f7qavoeawzxsndwbmoxnonqvsagp5ldn6pe
-valory/skills/apy_estimation_abci,bafybeiao7bm3kvl4cudlfqfeg7sjoey6amh3mudbxjj3h7z5j5bvfaiqqy
-valory/skills/apy_estimation_chained_abci,bafybeihkryqq2uj3gzxklr6bvlgywd7wfabw2tkptsxsdgrzlq4hpg5b6u
+valory/skills/abstract_round_abci,bafybeibantzb6stowqrjkijoljchxn7d6zyc6rmzvmx3tps3h2gduu3dpy
+valory/skills/apy_estimation_abci,bafybeiflvdg6fj5x5d5rxqimffe6gs2lhnfbzbwjpofeqm2x67oo7e5hci
+valory/skills/apy_estimation_chained_abci,bafybeievzkbzjvyd6ddmakhjtr4ooszkucpzziyw7yy6i4nvao5wufs3mq
 valory/skills/counter,bafybeigliv7i43y5ijjr3vsqypsnsbcusbfzjfhrpyjjeclokt45hepgry
 valory/skills/counter_client,bafybeiedl3s2xohuahn4wmoluy62ajirw7ipmjkgqemapqoqceu3lnv7z4
-valory/skills/hello_world_abci,bafybeig75jerm6yqudsrln3whgxip5wvw6u6cf33eq25zezxin6ddt4aoa
-valory/skills/liquidity_provision_abci,bafybeicdtqjhze466s77nfbhgxwuxirxcvkbv3jyhupvm3i7m3svh6nn3e
-valory/skills/liquidity_rebalancing_abci,bafybeicjxzg5thnt5zryjacxdxelxnbggax6zbfqtsqqw67f2owwzbypxi
-valory/skills/oracle_abci,bafybeihuocs4ze6kwkfrbr6eyyckoogpzcggqxgwti7p2qyi55p72hebzi
-valory/skills/oracle_deployment_abci,bafybeiab3zvejzn44l233ldw6wkoqza5nc4pa6qqist5hfprx5ikfzzxzu
-valory/skills/price_estimation_abci,bafybeiafzefuqpla2maovdlzwtp545im2ile7xm5zbfzviye6jlbz57uh4
-valory/skills/register_reset_abci,bafybeiet2izfwzso65n7knzebct4laohed7stc5iztyhwqqojutb625cjm
-valory/skills/registration_abci,bafybeig3dkrzo7m2wnw7t4n5nzcpxs4h5gcjvnwn7otrnimjtojk6y3wn4
-valory/skills/reset_pause_abci,bafybeifibdqxulkbe3qxyblnq6bzi4zwxdnage33tda6oiyvryficiskku
-valory/skills/safe_deployment_abci,bafybeif4aelv2jispmfadlywmreukow7ohhhkcqcyhqkus4axcedhhdifm
-valory/skills/simple_abci,bafybeiffi2n6bp7xdsmqy5rkmeq5knxucjda6ntbd3d5a3b43x4yaafrti
-valory/skills/test_abci,bafybeigtvismvmxp6jzk7fokmmoab5zecwxhkyhm62kcwhgtyuoryixhiy
-valory/skills/transaction_settlement_abci,bafybeia7gyr5ljmnyrmyydgc632pr2d2odxaq2hdsm7bbuc42gpfohkjba
+valory/skills/hello_world_abci,bafybeiblv6cydkgn2f2e4gtnaei3jpufv5tc4hofiwpzwhkray42ixjfva
+valory/skills/liquidity_provision_abci,bafybeicytbqzlrgivirjm26ptlogm4xsx5y4hv237piwigan4ann2nm4hu
+valory/skills/liquidity_rebalancing_abci,bafybeieeo67g2wcajb7kft4q5xfwaot2rtjuxetqzddfsjo4mukeqjdrii
+valory/skills/oracle_abci,bafybeibb2fn2ocjmg6noebspcajqvklrw3rqh75wjiamtupsxf7mk5tw7e
+valory/skills/oracle_deployment_abci,bafybeicrgypwb5xnpusxjuare725gbxkz6lczmxtmmltegg3d77fssaqaq
+valory/skills/price_estimation_abci,bafybeiedk76qidaxd7howyp6abf2xdwtp4i5wckq5vncrt3ihkqrgv235y
+valory/skills/register_reset_abci,bafybeia6sqqkmbznxxzypblfsihx3fa3srjmj3n763szbukn4kyp3bxuv4
+valory/skills/registration_abci,bafybeia624f7ic4uxdkouothup4pjjs46i6fcg6o5lwbnjdwbbc7vlncp4
+valory/skills/reset_pause_abci,bafybeies3onwza7ihrqd36l7daw5ruv4d3mturmaciinaigzafoxymu6km
+valory/skills/safe_deployment_abci,bafybeiahn6373j6rfykhsgju2fe4atvvku6nwuixwnad22jxw3vjhyftiy
+valory/skills/simple_abci,bafybeihw7wh27yd5mrickjmxgb5x4jmmqn4jey52ufy2bvfdqrq7sf2u5m
+valory/skills/test_abci,bafybeifs26jmc6r6is6k7xiqfpucx3nhvggoo4fndifj7vkpwyxuf3hxha
+valory/skills/transaction_settlement_abci,bafybeih2ml4l567s6bvnpguxo2wuya3v5kjc6ws2kzzhzk26azduducnfe
