+++ packages/valory/agents/apy_estimation_chained/aea-config.yaml
@@ -24,10 +24,10 @@ protocols:
 - valory/tendermint:0.1.0:bafybeihcnjhovvyyfbkuw5sjyfx2lfd4soeocfqzxz54g67333m6nk5gxq
 skills:
 - valory/abstract_abci:0.1.0:bafybeiam52ztcss46i3tgywuv4qmvddqgn7exmbt5kcicqpyip4kfeyap4
-- valory/abstract_round_abci:0.1.0:bafybeie4ux473o45jjo5cb2f7qavoeawzxsndwbmoxnonqvsagp5ldn6pe
-- valory/apy_estimation_abci:0.1.0:bafybeiao7bm3kvl4cudlfqfeg7sjoey6amh3mudbxjj3h7z5j5bvfaiqqy
-- valory/apy_estimation_chained_abci:0.1.0:bafybeihkryqq2uj3gzxklr6bvlgywd7wfabw2tkptsxsdgrzlq4hpg5b6u
-- valory/registration_abci:0.1.0:bafybeig3dkrzo7m2wnw7t4n5nzcpxs4h5gcjvnwn7otrnimjtojk6y3wn4
+- valory/abstract_round_abci:0.1.0:bafybeibantzb6stowqrjkijoljchxn7d6zyc6rmzvmx3tps3h2gduu3dpy
+- valory/apy_estimation_abci:0.1.0:bafybeiflvdg6fj5x5d5rxqimffe6gs2lhnfbzbwjpofeqm2x67oo7e5hci
+- valory/apy_estimation_chained_abci:0.1.0:bafybeievzkbzjvyd6ddmakhjtr4ooszkucpzziyw7yy6i4nvao5wufs3mq
+- valory/registration_abci:0.1.0:bafybeia624f7ic4uxdkouothup4pjjs46i6fcg6o5lwbnjdwbbc7vlncp4
 default_ledger: ethereum
 required_ledgers:
 - ethereum
