+++ packages/valory/agents/hello_world/aea-config.yaml
@@ -18,8 +18,8 @@ protocols:
 - valory/http:1.0.0:bafybeienmids6f5pcey2bnywmiup2faawqgmdhhz65qfciuhkntugztvpi
 skills:
 - valory/abstract_abci:0.1.0:bafybeiam52ztcss46i3tgywuv4qmvddqgn7exmbt5kcicqpyip4kfeyap4
-- valory/abstract_round_abci:0.1.0:bafybeie4ux473o45jjo5cb2f7qavoeawzxsndwbmoxnonqvsagp5ldn6pe
-- valory/hello_world_abci:0.1.0:bafybeig75jerm6yqudsrln3whgxip5wvw6u6cf33eq25zezxin6ddt4aoa
+- valory/abstract_round_abci:0.1.0:bafybeibantzb6stowqrjkijoljchxn7d6zyc6rmzvmx3tps3h2gduu3dpy
+- valory/hello_world_abci:0.1.0:bafybeiblv6cydkgn2f2e4gtnaei3jpufv5tc4hofiwpzwhkray42ixjfva
 default_ledger: ethereum
 required_ledgers:
 - ethereum
