+++ packages/valory/agents/oracle/aea-config.yaml
@@ -27,14 +27,14 @@ protocols:
 - valory/tendermint:0.1.0:bafybeihcnjhovvyyfbkuw5sjyfx2lfd4soeocfqzxz54g67333m6nk5gxq
 skills:
 - valory/abstract_abci:0.1.0:bafybeiam52ztcss46i3tgywuv4qmvddqgn7exmbt5kcicqpyip4kfeyap4
-- valory/abstract_round_abci:0.1.0:bafybeie4ux473o45jjo5cb2f7qavoeawzxsndwbmoxnonqvsagp5ldn6pe
-- valory/oracle_abci:0.1.0:bafybeihuocs4ze6kwkfrbr6eyyckoogpzcggqxgwti7p2qyi55p72hebzi
-- valory/oracle_deployment_abci:0.1.0:bafybeiab3zvejzn44l233ldw6wkoqza5nc4pa6qqist5hfprx5ikfzzxzu
-- valory/price_estimation_abci:0.1.0:bafybeiafzefuqpla2maovdlzwtp545im2ile7xm5zbfzviye6jlbz57uh4
-- valory/registration_abci:0.1.0:bafybeig3dkrzo7m2wnw7t4n5nzcpxs4h5gcjvnwn7otrnimjtojk6y3wn4
-- valory/reset_pause_abci:0.1.0:bafybeifibdqxulkbe3qxyblnq6bzi4zwxdnage33tda6oiyvryficiskku
-- valory/safe_deployment_abci:0.1.0:bafybeif4aelv2jispmfadlywmreukow7ohhhkcqcyhqkus4axcedhhdifm
-- valory/transaction_settlement_abci:0.1.0:bafybeia7gyr5ljmnyrmyydgc632pr2d2odxaq2hdsm7bbuc42gpfohkjba
+- valory/abstract_round_abci:0.1.0:bafybeibantzb6stowqrjkijoljchxn7d6zyc6rmzvmx3tps3h2gduu3dpy
+- valory/oracle_abci:0.1.0:bafybeibb2fn2ocjmg6noebspcajqvklrw3rqh75wjiamtupsxf7mk5tw7e
+- valory/oracle_deployment_abci:0.1.0:bafybeicrgypwb5xnpusxjuare725gbxkz6lczmxtmmltegg3d77fssaqaq
+- valory/price_estimation_abci:0.1.0:bafybeiedk76qidaxd7howyp6abf2xdwtp4i5wckq5vncrt3ihkqrgv235y
+- valory/registration_abci:0.1.0:bafybeia624f7ic4uxdkouothup4pjjs46i6fcg6o5lwbnjdwbbc7vlncp4
+- valory/reset_pause_abci:0.1.0:bafybeies3onwza7ihrqd36l7daw5ruv4d3mturmaciinaigzafoxymu6km
+- valory/safe_deployment_abci:0.1.0:bafybeiahn6373j6rfykhsgju2fe4atvvku6nwuixwnad22jxw3vjhyftiy
+- valory/transaction_settlement_abci:0.1.0:bafybeih2ml4l567s6bvnpguxo2wuya3v5kjc6ws2kzzhzk26azduducnfe
 default_ledger: ethereum
 required_ledgers:
 - ethereum
