+++ packages/valory/agents/simple_abci/aea-config.yaml
@@ -21,8 +21,8 @@ protocols:
 - valory/ledger_api:1.0.0:bafybeif2c3jt4tim3zbfq5c2lds7bs6qj5gvj2obk3pzdk3uy6zk2ivo3e
 skills:
 - valory/abstract_abci:0.1.0:bafybeiam52ztcss46i3tgywuv4qmvddqgn7exmbt5kcicqpyip4kfeyap4
-- valory/abstract_round_abci:0.1.0:bafybeie4ux473o45jjo5cb2f7qavoeawzxsndwbmoxnonqvsagp5ldn6pe
-- valory/simple_abci:0.1.0:bafybeiffi2n6bp7xdsmqy5rkmeq5knxucjda6ntbd3d5a3b43x4yaafrti
+- valory/abstract_round_abci:0.1.0:bafybeibantzb6stowqrjkijoljchxn7d6zyc6rmzvmx3tps3h2gduu3dpy
+- valory/simple_abci:0.1.0:bafybeihw7wh27yd5mrickjmxgb5x4jmmqn4jey52ufy2bvfdqrq7sf2u5m
 default_ledger: ethereum
 required_ledgers:
 - ethereum
