+++ packages/valory/services/apy_estimation_hardhat/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeibqyfur7wusrbfga66hcbauxvhs2g3j5yhknsxlvyh6mil57aajyu
 fingerprint_ignore_patterns: []
-agent: valory/apy_estimation:0.1.0:bafybeiep4mcz5n4r2uu4rco2gf4ouxptnmd6wt2ais6gbjmpteweziq4wy
+agent: valory/apy_estimation:0.1.0:bafybeihfufuc7gmkl3hvncgfaknx4xgz6klzigxrmm5jzbtincdfvw3bju
 number_of_agents: 4
 ---
 public_id: valory/apy_estimation_abci:0.1.0
