+++ packages/valory/services/hello_world/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeiapubcoersqnsnh3acia5hd7otzt7kjxekr6gkbrlumv6tkajl6jm
 fingerprint_ignore_patterns: []
-agent: valory/hello_world:0.1.0:bafybeibotci3wyi6nyduztw6nhcosgci7brs2z3uzvs3fph7xxmcjm2xje
+agent: valory/hello_world:0.1.0:bafybeigjyz4pd7ylyxxpryadz7y5q7idxxamm7xh6rh2gtqz2tt276ii4e
 number_of_agents: 4
 ---
 benchmark_persistence_params:
