+++ packages/valory/services/oracle_goerli/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeig4wu7inb6ibwfgged3djyc7emopcg5n7d7gwrlotne53pp3jfgq4
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeiexc53bm7af643wcbdwa47dipfim3lapymru2emw76b3ayv4baxma
+agent: valory/oracle:0.1.0:bafybeih2qdphkai3flbuywwmwmghkcbsfqephlszfa7vdbifnwtzm3a7ne
 number_of_agents: 4
 ---
 setup_args:
