+++ packages/valory/services/oracle_ropsten_cluster/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeidtphhzkcp32sikeqoo4ct57z6kw3yhxkebby3unuwepqdvwgfkga
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeiexc53bm7af643wcbdwa47dipfim3lapymru2emw76b3ayv4baxma
+agent: valory/oracle:0.1.0:bafybeih2qdphkai3flbuywwmwmghkcbsfqephlszfa7vdbifnwtzm3a7ne
 number_of_agents: 4
 ---
 setup_args:
