+++ packages/valory/services/register_reset/service.yaml
@@ -1,7 +1,7 @@
 name: register_reset
 author: valory
 version: 0.1.0
-agent: valory/register_reset:0.1.0:bafybeiguas2i2xjsh5ngtkgfofzng4tyjonxhvxplzqdozgm6votmk4ape
+agent: valory/register_reset:0.1.0:bafybeicdbjj7gknew3efdvus64w4rq33mns4qzipejwjvqxpd4jculxkpm
 number_of_agents: 4
 description: Test and debug tendermint reset mechanism.
 aea_version: '>=1.0.0, <2.0.0'
