+++ packages/valory/services/simple_abci/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeia3zelg4urfp2g4srt6p3prza5et2hcwk57gsntiabut6ljs6yihe
 fingerprint_ignore_patterns: []
-agent: valory/simple_abci:0.1.0:bafybeiei4nte44ryfyjbpahimzir4xtepqss7x42kcq5iljzg5vsesnzi4
+agent: valory/simple_abci:0.1.0:bafybeieuog75diggn7kxa757nyz3jyhjs7qy75pmzvfxc2vi35b6qa3hcq
 number_of_agents: 4
 ---
 benchmark_persistence_params:
