+++ packages/valory/skills/abstract_round_abci/io_/paths.py
