+++ packages/valory/skills/abstract_round_abci/skill.yaml
@@ -10,16 +10,16 @@ fingerprint:
   __init__.py: bafybeiaxhdfjmhbstyi5oqeu7wzst44snhbyoggrpusgs2c3tn3cluzhpm
   abci_app_chain.py: bafybeihb3ded23gvjrn426qptkbnkge5cewqm7ywi4j6ebmo3j66ss5wii
   base.py: bafybeicbdo74prebsmczkpmpj7lqs2cqkgjh6uwsvas7cztabxfttain6y
-  behaviour_utils.py: bafybeiavg4eq7zc7t3qrictsvoqoz4si2qaxt5njinsgundhxhafmhiq24
+  behaviour_utils.py: bafybeibdrj3fsfqghol5ay5u32isvjwgxwm656op6fulubt4qutswpqtui
   behaviours.py: bafybeif3knc6efobddgkh3xc3yqwgtlk7z4zflftd7sic5j3at5rrhsute
   common.py: bafybeicz7feaxvwftwi7qe2x4g4f6edwsh434kr4duebafbwphqyxzuyka
   dialogues.py: bafybeibvnmcn7kofol2h6owljhde4llfc3kwwkb5helxsgsy47mokhwb4i
   handlers.py: bafybeidammqyvz2wvqirr3it4tfy7vbe5kxybwkz5o3y4ogo3wrbaqwcmy
-  io/__init__.py: bafybeiattfgecxl5y276etqe5jiuhqk3grcov6qky5cgokstmwhrbaodri
-  io/ipfs.py: bafybeigmsaoxdcpg5qvsxjkhilpfvcdmpsb2lie4fu5oqc2ows6i7pifwa
-  io/load.py: bafybeib63wgk7wcghrvldqn6dpurkw6o7q3eohg3dpinnf7rhw4qwo2j7q
-  io/paths.py: bafybeicfno2l4vwtmjcm3rzpp6tqi3xlkof47pypf5teecad22d44u2ple
-  io/store.py: bafybeiee7uhs372q3kxd5s3itufb3ws24ey74my2jpnivit3kboon6xsaq
+  io_/__init__.py: bafybeiattfgecxl5y276etqe5jiuhqk3grcov6qky5cgokstmwhrbaodri
+  io_/ipfs.py: bafybeic2lil23fk77wk3i6jrx47fdmc5wxr4dl2xbjjomgtfkz7yesz4eq
+  io_/load.py: bafybeiamag3kolsuv75skjgzfyosip3qqpdqzadet6niqej7x4wdostgdm
+  io_/paths.py: bafybeicfno2l4vwtmjcm3rzpp6tqi3xlkof47pypf5teecad22d44u2ple
+  io_/store.py: bafybeiczvoewpvhlblwkewsx4b27urji73vgssrhxkkpqf4a6eesmqp62m
   models.py: bafybeifsna4hnkhgqegmhzh2uh62q6ikplvoqcsujnnvpol3wqozdhpcn4
   serializer.py: bafybeiezjdtp33nub7ual5ibt4vi7dj3ilmyw7wqq374ctub5oo7wgp7mm
   test_tools/__init__.py: bafybeigicwqumbl3jorusndtfa6ag7lnf3saa6cbqn7fazur4n6nxmsdva
