+++ packages/valory/skills/apy_estimation_abci/behaviours.py
@@ -47,7 +47,7 @@ from packages.valory.skills.abstract_round_abci.behaviours import (
     AbstractRoundBehaviour,
     BaseBehaviour,
 )
-from packages.valory.skills.abstract_round_abci.io.load import SupportedFiletype
+from packages.valory.skills.abstract_round_abci.io_.load import SupportedFiletype
 from packages.valory.skills.abstract_round_abci.models import ApiSpecs
 from packages.valory.skills.abstract_round_abci.utils import VerifyDrand
 from packages.valory.skills.apy_estimation_abci.ml.forecasting import (
@@ -114,7 +114,7 @@ from packages.valory.skills.apy_estimation_abci.tools.general import (
     sec_to_unit,
     unit_amount_from_sec,
 )
-from packages.valory.skills.apy_estimation_abci.tools.io import load_hist
+from packages.valory.skills.apy_estimation_abci.tools.io_ import load_hist
 from packages.valory.skills.apy_estimation_abci.tools.queries import (
     block_from_number_q,
     block_from_timestamp_q,
