+++ tests/test_skills/test_apy_estimation_abci/test_behaviours.py
@@ -59,7 +59,7 @@ from packages.valory.skills.abstract_round_abci.behaviour_utils import (
     BaseBehaviour,
     IPFSBehaviour,
 )
-from packages.valory.skills.abstract_round_abci.io.store import SupportedFiletype
+from packages.valory.skills.abstract_round_abci.io_.store import SupportedFiletype
 from packages.valory.skills.abstract_round_abci.models import ApiSpecs, BenchmarkTool
 from packages.valory.skills.abstract_round_abci.test_tools.base import (
     FSMBehaviourBaseCase,
